+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+Renamed:
+- `SD-Submit` to `SD Apply` in the UI and logs
+- `SD-Connect` to `SD Connect` in the UI and logs
+
 ## [v1.0.0] - 2022-03-04
 
 ### Changed